resolved all nits
