+++ constants/constants.ts
@@ -1,0 +1,3 @@
+export const service_id = "service_o683bxk";
+export const template_id = "template_2db4zzm";
+export const user_id = "user_jlv6FzwqKDExG2SkcZmga";
