+++ firebase/firestore/admin.ts
@@ -27,6 +27,7 @@ export const getAdmin = async (adminId: string): Promise<Admin> => {
     const adminData = doc.data() as Admin;
     // enables date objects to be serializable during SSR
     adminData.created_at = adminData.created_at.toDate().toString();
+    adminData.last_active = adminData.last_active.toDate().toString();
     return adminData;
   } catch (e) {
     console.error(e);
