+++ types/schema.ts
@@ -18,6 +18,7 @@ export type Family = {
 
 export type Admin = {
   created_at: Date | Timestamp | string;
+  last_active: Date | Timestamp | string;
   email: string;
   full_name: string;
 };
